requirements.txt
customtkinter
tkinterdnd2
pillow
openai
openpyxl
python-docx
pymupdf